streaming_config_params/requirements.txt
azure-eventhub
delta-spark
pyarrow
typing_extensions
pandas
zstandard
msgpack